requirements.txt
absl-py==0.5.0
astor==0.7.1
awscli==1.16.33
backports.weakref==1.0.post1
botocore==1.12.40
colorama==0.3.9
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
jmespath==0.9.3
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
mock==2.0.0
numpy==1.15.2
pbr==4.3.0
protobuf==3.6.1
pyasn1==0.4.4
python-dateutil==2.7.3
PyYAML==3.13
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.14.1
youtube-dl==2019.2.18
